refactoring; opening CVSX from 'Open File' menu
